requirements.txt
pandas
streamlit
scikit-learn==1.2.2
joblib==1.4.2
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
numpy==1.25.2
